提交代码

+++ src/main/java/com/cgs/entity/index/KDJItem.java
@@ -5,8 +5,8 @@ import lombok.Data;
 @Data
 public class KDJItem {
     private String stockId;
-    private Double kValue;
-    private Double dValue;
-    private Double jValue;
+    private Integer kValue;
+    private Integer dValue;
+    private Integer jValue;
     private String date;
 }
